src/interfaces/pureArchAuth/logout.go
package pureArchAuth

import (
	"net/http"
	"server/src/domain/repository"
	"time"
)

func (a *Authenticate) Logout(w http.ResponseWriter, r *http.Request) {
	cookie, err := r.Cookie("session_id")
	if err == http.ErrNoCookie {
		w.WriteHeader(http.StatusFound)
		return
	}

	if err = a.auth.DeleteAuth(&repository.AccessDetails{Value: cookie.Value}); err != nil {
		a.log.Print(err)
		w.WriteHeader(http.StatusInternalServerError)
		return
	}

	newCookie, err := a.cookie.CreateCookie()
	if err != nil {
		a.log.Print(err)
		w.WriteHeader(http.StatusInternalServerError)
		return
	}

	newCookie.Expires = time.Date(1973, 1, 1, 0, 0, 0, 0, time.UTC)
	newCookie.Value = ""
	http.SetCookie(w, &newCookie)
	w.WriteHeader(http.StatusFound)
}
